+++ tap_exacttarget/streams/__init__.py
@@ -1,19 +1,19 @@
 from .campaigns import Campaigns
 from .content_area import ContentArea
 from .dataextensionobjects import DataExtensionObjectFt, DataExtensionObjectInc  # noqa: F401
-from .emails import Email
+from .email import Email
 from .event_bounce import BounceEvent
 from .event_click import ClickEvent
 from .event_notsent import NotSentEvent
 from .event_open import OpenEvent
 from .event_sent import SentEvent
 from .event_unsub import UnsubEvent
-from .folders import DataFolder
+from .datafolder import DataFolder
 from .list_send import ListSend
-from .list_subscribers import ListSubscribe
-from .lists import ETList
-from .sends import Sends
-from .subscribers import Subscribers
+from .list_subscribers import ListSubscribers
+from .list import ETList
+from .send import Sends
+from .subscriber import Subscribers
 
 STREAMS = {
     Campaigns.tap_stream_id: Campaigns,
@@ -29,6 +29,6 @@ STREAMS = {
     NotSentEvent.tap_stream_id: NotSentEvent,
     UnsubEvent.tap_stream_id: UnsubEvent,
     ListSend.tap_stream_id: ListSend,
-    ListSubscribe.tap_stream_id: ListSubscribe,
+    ListSubscribers.tap_stream_id: ListSubscribers,
     Subscribers.tap_stream_id: Subscribers,
 }
